requirements.txt
# Core App & Data
streamlit
pandas
requests

# Machine Learning
scikit-learn
tensorflow
joblib
Pillow

# APIs & Accessibility
owslib
gTTS
pyproj